pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mrtrix-rish"
version = "0.1.0"
description = "MRtrix3-native RISH harmonization for multi-site diffusion MRI"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Arush", email = "arush.py@gmail.com"}
]
keywords = [
    "neuroimaging",
    "diffusion-mri",
    "harmonization",
    "mrtrix3",
    "spherical-harmonics"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
]
requires-python = ">=3.9"
dependencies = [
    "numpy>=1.20",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
qc = [
    "nibabel>=4.0",
    "matplotlib>=3.5",
    "jinja2>=3.0",
]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "ruff>=0.1",
    "mypy>=1.0",
]
all = [
    "mrtrix-rish[qc,dev]",
]

[project.scripts]
mrtrix-rish = "src.cli.main:main"

[project.urls]
Homepage = "https://github.com/arush/mrtrix-rish"
Documentation = "https://mrtrix-rish.readthedocs.io"
Repository = "https://github.com/arush/mrtrix-rish"
Issues = "https://github.com/arush/mrtrix-rish/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.black]
line-length = 88
target-version = ["py39", "py310", "py311", "py312"]

[tool.ruff]
line-length = 88
select = ["E", "F", "W", "I", "UP", "B", "C4"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
addopts = "-v --cov=src --cov-report=term-missing"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
